+++ action-confirmation-auditor/scripts/Start-ActionConfirmationValidationRunbook.ps1
@@ -273,8 +273,8 @@ try {
     $previousAgentMap = @{}
 
     try {
-        $prevScanUri = "$baseUrl/api/data/v9.2/fsi_actionscanruns?" +
-            "`$orderby=fsi_scantime desc&`$top=1&" +
+        $prevScanUri = "$baseUrl/api/data/v9.2/fsi_actionscanrun?" +
+            "`$orderby=fsi_validationtime desc&`$top=1&" +
             "`$select=fsi_runid,fsi_summaryjson"
 
         $prevScanResponse = Invoke-RestMethod -Uri $prevScanUri -Method Get -Headers $driftHeaders -ErrorAction Stop